src/pyproject.toml
[project]
name = "quoteoftheday"
version = "1.0.0"
description = ""
dependencies = [
    "azure-monitor-opentelemetry",
    "Flask==2.3.2",
    "SQLAlchemy==2.0.17",
    "psycopg2-binary==2.9.6",
    "Flask-Migrate==4.0.4",
    "Flask-SQLAlchemy==3.1.1",
    "azure-identity==1.12.0",
    "azure-appconfiguration-provider==2.0.0b1",
    "featuremanagement[AzureMonitor]==2.0.0b1",
    "flask-login",
    "flask_sqlalchemy",
    "flask_bcrypt"
    ]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
